+++ NEWS.md
@@ -8,6 +8,7 @@
 * Introduce `facet_var` drop down in `tm_g_gh_scatterplot` and `tm_g_gh_correlationplot`.
 * Replace `plot_with_height` module with new `plot_with_settings` module.
 * Allow to set a minimum number of observations to render a node in `tm_g_gh_lineplot` and allow to change the font size in the table. Also the previously named argument `font_size` is now `plot_font_size` to distinguish it from the table font size argument.
+* Add option to customize symbol type, line type and color in `tm_g_gh_lineplot`.
 * issues with no news:
 
 # teal.goshawk 0.1.5